#1691 - test refactoring
